% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logistic2.R
\name{Logreg}
\alias{Logreg}
\title{Binomial Logistic regression}
\usage{
Logreg(X, y, maxit = 5000)
}
\arguments{
\item{X}{input matrix, of dimension n (sample number) by p (variable number);
each row is an observation vector.
Can be in sparse matrix format
(inherit from class "sparseMatrix" as in package Matrix;}

\item{y}{response variable.
Since we aim at binomial distribution, it should be either a factor with two levels,
 or a two-column matrix (every row is the probability of two class.)}

\item{maxit}{the Maximum number of iterations when we use optimization to estimate the parameeter;
 default is 10^5.

 @return a data frame which contains the result of this logistic regression.

 @param x the regression result, a length p+1 vector (include the intercept)

 @param loss the record of loss in iterations, which can help users to check the convergence

 @param Train_Acc the accuracy of the train set, defined as (number of right prediction divided by sample size)

 @export}
}
\description{
Fit a generalized linear model via penalized maximum likelihood.
The regularization path is computed for the lasso or elasticnet penalty
at a grid of values for the regularization parameter lambda.
Can deal with all shapes of data,including very large sparse data matrices.
}
