% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_predict.R
\name{My_predict}
\alias{My_predict}
\title{Use the model generated by "LogReg" to predict new test data}
\usage{
My_predict(fit, newx)
}
\arguments{
\item{fit}{the model generated by "LogReg".
Since we aim at binomial distribution, it should be either a factor with two levels,
 or a two-column matrix (every row is the probability of two class.)
 @return the prediction for every row of test data newx. And the label will be the same as that of training data.
  @export}

\item{newx}{input matrix, of dimension n (sample number) by p (variable number);
each row is an observation vector.}
}
\description{
Use the model generated by "LogReg" to predict new test data
}
